+++ CHANGELOG.md
@@ -46,6 +46,8 @@ For each minor and major version, there is a corresponding [milestone on GitHub]
     }
     ```
 
+  - It is possible to customize the `ValidationMessageDetails` now, which is used to provide the Langium-specific validation properties in Typir-Langium, e.g. to support code actions for validation issues reported by Typir (see `lox-type-checking-operators.test.ts` for LOX).
+
 ### Breaking changes
 
 - Typir-Langium: `LangiumLanguageNodeInferenceCaching` and `DefaultLangiumTypeCreator` use the `TypirLangiumServices` parameter to retrieve the `LangiumSharedCoreServices` now (#78).